feat(imx8mp): allow access to UART1 from A53 only
TF-A sets the Peripheral Domain Access Permissions (RDC_PDAPn), where
access to UART3/4 are limited to M7 and access to UART2 is limited to A53
on i.MX8MP platform.

Allow read and write access to UART1 from Domain 0 (A53), as it is used
by bootloaders and Linux OS in some designs as a debug UART alternative
to NXP EVK setup.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>