fix(imx8m): allow write access to snvs LPSR register to reset SPO
Allow write access to the snvs registers - namely the Low-Power Status Register
(LPSR) to be able to reset the interrupt flag BIT(18) = SPO, which is required
for the linux-kernel drivers/input/keyboard/snvs_pwrkey to be able/allowed to
reset the interrupt flag and continue handling the BTN status (via polling).

The LPSR register used to be 'non privileged' on the imx6, for which the kernel
code was probably written, but changed to a priviledged register with the newer
imxs; now on a secured ic, the interrupt keeps firing with no means of resetting
it, because the access to the priviledged registers is disabled.

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>