feat(lfa): implement LFA_SVC_GET_INVENTORY SMC
This implementation primarily retrieves the version, activation flags
such as activation capable, activation pending, and the UUID, all
corresponding to the given fw_seq_id. To achieve this, several platform
APIs were created to fetch the firmware version of the given fw_seq_id
and determine if an update is available, indicating that it is pending
activation.

Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>
Change-Id: I6e702a8b313386f5668027386845ff2dae9328e7