feat(per-cpu): introduce per-cpu framework accessors and definers
introduce per-cpu framework definers and accessors for allocation and
access of per-CPU objects. The accessors support "this-cpu" variants for
access for per-cpu object on calling cpu as well as "for-cpu" variants
for access of per-cpu objects on any cpu.

Additionally, the framework supports NUMA-aware allocation, allowing the
per-cpu data to be distributed across different memory nodes. This
enables the system to allocate per-cpu data on memory nodes closest to
the respective CPU, optimising memory access and performance.

Change-Id: I105eceeedaa679336fcb120fccad82060fec6c99
Signed-off-by: Sammit Joshi <sammit.joshi@arm.com>
Signed-off-by: Rohit Mathew <rohit.mathew@arm.com>