feat(lfa): xlat changes for BL31 LFA
This patch adds xlat library changes needed to support live
activation of BL31 images. During a live firmware activation,
the running BL31 image must be overwritten, which requires read
only memory to be marked as writable.

This is done by adding separate functions to the Xlat library
with reduced restrictions that are only compiled in when
LFA_SUPPORT=1.

Signed-off-by: John Powell <john.powell@arm.com>
Change-Id: I3956542b7978133dddcb0d1ace5512b6cfbfedc0